requirements.txt
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==2.3.1
supabase==2.7.2
Unidecode==1.3.8
